feat/refac: 슬라이더 영상 지원 추가, 중복 코드 통합 및 영상 로직 커스텀 훅으로 분리 - 멀티뷰, 단일뷰 상태에서 중복되는 슬라이더 코드 renderSlider로 통합 - logList, videoViewer, imageModal에서 공통으로 사용되는 영상 함수 커스텀 훅으로 분리
