+++ service/src/pages/mainPage/components/log/LogList.scss
@@ -209,7 +209,7 @@
     background-color: #2e2e2e;
     padding: 50px 12px;
     text-align: center;
-    font-size: 14px;
+    font-size: 12px;
     width: 100%;
     @include mobile {
         padding: 20px 12px;
